+++ src/app/shared/dto/users/account/CreatedServiceProviderAccountDTO.ts
@@ -1,9 +1,8 @@
-import {PersonType} from '../person/PersonType.model';
-import {RegistrationRequestDTO} from '../registrationRequest/RegistrationRequestDTO.model';
+import {PersonType} from '../../../model/PersonType.model';
 import {SimpleServiceProviderDTO} from '../serviceProvider/SimpleServiceProviderDTO.model';
+import {RegistrationRequestDTO} from '../../registrationRequest/RegistrationRequestDTO.model';
 
-
-export interface CreatedServiceProviderAccountDTOModel{
+export interface CreatedServiceProviderAccountDTO {
   id: string;
   email: string;
   isActive: boolean;
